// Frontend/src/app/models/admin/real-time.ts
export class RealTime {
    id?: Number;
    last_notified_loan?: Number;
    last_notification?: Number;
    last_notification_admin?: Number;
    user_id?: Number;
    region_id?: Number;
    is_saved?: Boolean;
    approver_approve?: Boolean;
    approver_reject?: Boolean;
    maker_edit?: Boolean;
    maker_disburse?: Boolean;
    notify?: Boolean;
    isAdmin?: Boolean;
}